refactor output methods
